+++ axelor-production/src/main/java/com/axelor/apps/production/service/costsheet/CostSheetLineServiceImpl.java
@@ -17,6 +17,7 @@
  */
 package com.axelor.apps.production.service.costsheet;
 
+import com.axelor.apps.base.db.Company;
 import com.axelor.apps.base.db.Product;
 import com.axelor.apps.base.db.Unit;
 import com.axelor.apps.base.db.repo.ProductRepository;
@@ -29,6 +30,7 @@ import com.axelor.apps.production.db.WorkCenter;
 import com.axelor.apps.production.db.repo.CostSheetGroupRepository;
 import com.axelor.apps.production.db.repo.CostSheetLineRepository;
 import com.axelor.apps.production.service.app.AppProductionService;
+import com.axelor.apps.stock.service.WeightedAveragePriceService;
 import com.axelor.exception.AxelorException;
 import com.beust.jcommander.internal.Lists;
 import com.google.inject.Inject;
@@ -47,17 +49,20 @@ public class CostSheetLineServiceImpl implements CostSheetLineService {
   protected CostSheetGroupRepository costSheetGroupRepository;
   protected UnitConversionService unitConversionService;
   protected UnitRepository unitRepo;
+  protected WeightedAveragePriceService weightedAveragePriceService;
 
   @Inject
   public CostSheetLineServiceImpl(
       AppProductionService appProductionService,
       CostSheetGroupRepository costSheetGroupRepository,
       UnitConversionService unitConversionService,
-      UnitRepository unitRepo) {
+      UnitRepository unitRepo,
+      WeightedAveragePriceService weightedAveragePriceService) {
     this.appProductionService = appProductionService;
     this.costSheetGroupRepository = costSheetGroupRepository;
     this.unitConversionService = unitConversionService;
     this.unitRepo = unitRepo;
+    this.weightedAveragePriceService = weightedAveragePriceService;
   }
 
   public CostSheetLine createCostSheetLine(
@@ -157,42 +162,41 @@ public class CostSheetLineServiceImpl implements CostSheetLineService {
   }
 
   public CostSheetLine createConsumedProductCostSheetLine(
+      Company company,
       Product product,
       Unit unit,
       int bomLevel,
       CostSheetLine parentCostSheetLine,
-      BigDecimal consumptionQty)
+      BigDecimal consumptionQty,
+      int origin)
       throws AxelorException {
 
-    BigDecimal price;
+    Product parentProduct = parentCostSheetLine.getProduct();
 
-    if ((product.getProductSubTypeSelect().equals(ProductRepository.PRODUCT_SUB_TYPE_COMPONENT)
-            || product
-                .getProductSubTypeSelect()
-                .equals(ProductRepository.PRODUCT_SUB_TYPE_SEMI_FINISHED_PRODUCT))
-        && parentCostSheetLine
-            .getProduct()
-            .getRealOrEstimatedPriceSelect()
-            .equals(ProductRepository.PRICE_METHOD_REAL)
-        && parentCostSheetLine
-            .getProduct()
-            .getComponentsValuationMethod()
-            .equals(ProductRepository.COMPONENTS_VALUATION_METHOD_AVERAGE)) {
-      price = product.getAvgPrice();
-    } else {
-      price = product.getCostPrice();
+    BigDecimal costPrice = null;
+    switch (origin) {
+      case ORIGIN_BILL_OF_MATERIAL:
+        costPrice =
+            this.getComponentCostPrice(
+                parentProduct, parentProduct.getBomCompValuMethodSelect(), company);
+        break;
+
+      case ORIGIN_MANUF_ORDER:
+        costPrice =
+            this.getComponentCostPrice(
+                product, parentProduct.getManufOrderCompValuMethodSelect(), company);
+        break;
+
+      default:
+        costPrice = BigDecimal.ZERO;
     }
 
-    if (price.compareTo(BigDecimal.ZERO) == 0) {
-      price = product.getPurchasePrice();
-    }
-
-    BigDecimal costPrice =
+    costPrice =
         unitConversionService
             .convert(
                 unit,
                 product.getUnit(),
-                price,
+                costPrice,
                 appProductionService.getNbDecimalDigitForUnitPrice(),
                 product)
             .multiply(consumptionQty);
@@ -209,6 +213,29 @@ public class CostSheetLineServiceImpl implements CostSheetLineService {
         unit,
         null,
         parentCostSheetLine);
+  }
+
+  protected BigDecimal getComponentCostPrice(
+      Product product, int componentsValuationMethod, Company company) {
+
+    BigDecimal price = null;
+
+    if (componentsValuationMethod == ProductRepository.COMPONENTS_VALUATION_METHOD_AVERAGE) {
+      price = weightedAveragePriceService.computeAvgPriceForCompany(product, company);
+    } else if (componentsValuationMethod == ProductRepository.COMPONENTS_VALUATION_METHOD_COST) {
+      price = product.getCostPrice();
+    }
+    if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
+      price = product.getCostPrice();
+    }
+    if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
+      price = weightedAveragePriceService.computeAvgPriceForCompany(product, company);
+    }
+    if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
+      price = product.getPurchasePrice();
+    }
+
+    return price;
   }
 
   public CostSheetLine createConsumedProductWasteCostSheetLine(
@@ -264,14 +291,7 @@ public class CostSheetLineServiceImpl implements CostSheetLineService {
       Unit unit) {
 
     return this.createWorkCenterCostSheetLine(
-        workCenter,
-        priority,
-        bomLevel,
-        parentCostSheetLine,
-        consumptionQty,
-        costPrice,
-        unit,
-        prodHumanResource.getCostSheetGroup());
+        workCenter, priority, bomLevel, parentCostSheetLine, consumptionQty, costPrice, unit, null);
   }
 
   public CostSheetLine createWorkCenterMachineCostSheetLine(